+++ packages/circuit/quri_parts/circuit/__init__.py
@@ -70,37 +70,6 @@ from .parameter_mapping import (
     ParameterMapping,
     ParameterOrLinearFunction,
 )
-from .transpile import (
-    CircuitTranspiler,
-    CliffordApproximationTranspiler,
-    CNOT2CZHTranspiler,
-    CZ2CNOTHTranspiler,
-    H2RZSqrtXTranspiler,
-    IdentityInsertionTranspiler,
-    PauliDecomposeTranspiler,
-    PauliRotationDecomposeTranspiler,
-    RX2RZSqrtXTranspiler,
-    RY2RZSqrtXTranspiler,
-    RZSetTranspiler,
-    S2RZTranspiler,
-    Sdag2RZTranspiler,
-    SequentialTranspiler,
-    SqrtX2RZHTranspiler,
-    SqrtXdag2RZSqrtXTranspiler,
-    SqrtY2RZSqrtXTranspiler,
-    SqrtYdag2RZSqrtXTranspiler,
-    SWAP2CNOTTranspiler,
-    T2RZTranspiler,
-    Tdag2RZTranspiler,
-    U1ToRZTranspiler,
-    U2ToRZSqrtXTranspiler,
-    U3ToRZSqrtXTranspiler,
-    X2HZTranspiler,
-    X2SqrtXTranspiler,
-    Y2RZXTranspiler,
-    Z2HXTranspiler,
-    Z2RZTranspiler,
-)
 
 #: A placeholder representing a constant term.
 CONST = CONST
@@ -163,35 +132,6 @@ __all__ = [
     "LinearMappedUnboundParametricQuantumCircuitBase",
     "LinearMappedUnboundParametricQuantumCircuit",
     "ImmutableLinearMappedUnboundParametricQuantumCircuit",
-    "CircuitTranspiler",
-    "SequentialTranspiler",
-    "PauliDecomposeTranspiler",
-    "PauliRotationDecomposeTranspiler",
-    "CliffordApproximationTranspiler",
-    "CNOT2CZHTranspiler",
-    "CZ2CNOTHTranspiler",
-    "SWAP2CNOTTranspiler",
-    "Z2HXTranspiler",
-    "X2HZTranspiler",
-    "X2SqrtXTranspiler",
-    "SqrtX2RZHTranspiler",
-    "H2RZSqrtXTranspiler",
-    "Y2RZXTranspiler",
-    "Z2RZTranspiler",
-    "SqrtXdag2RZSqrtXTranspiler",
-    "SqrtY2RZSqrtXTranspiler",
-    "SqrtYdag2RZSqrtXTranspiler",
-    "S2RZTranspiler",
-    "Sdag2RZTranspiler",
-    "T2RZTranspiler",
-    "Tdag2RZTranspiler",
-    "RX2RZSqrtXTranspiler",
-    "RY2RZSqrtXTranspiler",
-    "U1ToRZTranspiler",
-    "U2ToRZSqrtXTranspiler",
-    "U3ToRZSqrtXTranspiler",
-    "IdentityInsertionTranspiler",
-    "RZSetTranspiler",
     "inverse_gate",
     "is_clifford",
 ]
