+++ packages/circuit/tests/circuit/transpile/test_gate_kind_decomposer.py
@@ -20,115 +20,49 @@ from quri_parts.circuit import (
     U1,
     U2,
     U3,
-    H,
-    Identity,
-    Pauli,
-    PauliRotation,
-    QuantumCircuit,
-    S,
-    Sdag,
-    SqrtX,
-    SqrtXdag,
-    SqrtY,
-    SqrtYdag,
-    T,
-    Tdag,
-    X,
-    Y,
-    Z,
-)
-from quri_parts.circuit.transpile import (
     CliffordApproximationTranspiler,
     CZ2CNOTHTranspiler,
+    H,
     H2RZSqrtXTranspiler,
+    Identity,
     IdentityInsertionTranspiler,
+    Pauli,
     PauliDecomposeTranspiler,
+    PauliRotation,
     PauliRotationDecomposeTranspiler,
+    QuantumCircuit,
     RX2RZSqrtXTranspiler,
     RY2RZSqrtXTranspiler,
     RZSetTranspiler,
+    S,
     S2RZTranspiler,
+    Sdag,
     Sdag2RZTranspiler,
+    SqrtX,
     SqrtX2RZHTranspiler,
+    SqrtXdag,
     SqrtXdag2RZSqrtXTranspiler,
+    SqrtY,
     SqrtY2RZSqrtXTranspiler,
+    SqrtYdag,
     SqrtYdag2RZSqrtXTranspiler,
     SWAP2CNOTTranspiler,
+    T,
     T2RZTranspiler,
+    Tdag,
     Tdag2RZTranspiler,
     U1ToRZTranspiler,
     U2ToRZSqrtXTranspiler,
     U3ToRZSqrtXTranspiler,
+    X,
     X2HZTranspiler,
     X2SqrtXTranspiler,
+    Y,
     Y2RZXTranspiler,
+    Z,
     Z2HXTranspiler,
     Z2RZTranspiler,
 )
-
-
-class TestIdentityInsertion:
-    def test_fill_void(self) -> None:
-        circuit = QuantumCircuit(3)
-        circuit.extend([H(0), CNOT(0, 2), X(2)])
-        transpiled = IdentityInsertionTranspiler()(circuit)
-
-        expect = QuantumCircuit(3)
-        expect.extend([H(0), CNOT(0, 2), X(2), Identity(1)])
-
-        assert transpiled.gates == expect.gates
-
-    def test_fill_empty(self) -> None:
-        circuit = QuantumCircuit(2)
-        transpiled = IdentityInsertionTranspiler()(circuit)
-
-        expect = QuantumCircuit(2)
-        expect.extend([Identity(0), Identity(1)])
-
-        assert transpiled.gates == expect.gates
-
-    def test_no_change(self) -> None:
-        circuit = QuantumCircuit(2)
-        circuit.extend([H(0), SWAP(1, 0)])
-        transpiled = IdentityInsertionTranspiler()(circuit)
-
-        expect = QuantumCircuit(2)
-        expect.extend([H(0), SWAP(1, 0)])
-
-        assert transpiled.gates == expect.gates
-
-
-class TestMultiPauliDecompose:
-    def test_pauli_decompose(self) -> None:
-        circuit = QuantumCircuit(3)
-        circuit.add_gate(Pauli((1, 0, 2), (3, 1, 2)))
-        transpiled = PauliDecomposeTranspiler()(circuit)
-
-        expect = QuantumCircuit(3)
-        expect.extend([Z(1), X(0), Y(2)])
-
-        assert transpiled.gates == expect.gates
-
-    def test_pauli_rotation_decompose(self) -> None:
-        circuit = QuantumCircuit(3)
-        circuit.add_gate(PauliRotation((0, 2, 1), (2, 3, 1), np.pi / 4.0))
-        transpiled = PauliRotationDecomposeTranspiler()(circuit)
-
-        expect = QuantumCircuit(3)
-        gates = [
-            RX(0, np.pi / 2.0),
-            H(1),
-            CNOT(2, 0),
-            CNOT(1, 0),
-            RZ(0, np.pi / 4.0),
-            CNOT(2, 0),
-            CNOT(1, 0),
-            RX(0, -np.pi / 2.0),
-            H(1),
-        ]
-        expect.extend(gates)
-
-        assert transpiled.gates == expect.gates
 
 
 class TestFTQCSetTranspile:
@@ -374,51 +308,6 @@ class TestRZSetTranspile:
                 RZ(0, theta + np.pi),
                 SqrtX(0),
                 RZ(0, phi + 3.0 * np.pi),
-            ]
-        )
-
-        assert transpiled.gates == expect.gates
-
-    def test_cliffordapproximation_transpiler(self) -> None:
-
-        gate_list = [
-            H(0),
-            T(1),
-            RY(0, 1 * np.pi / 2),
-            RX(1, 1.4 * np.pi / 2),
-            U2(1, 0.5 * np.pi / 2, 1.7 * np.pi / 2),
-            U3(3, 0.5 * np.pi / 2, 1.5 * np.pi / 2, 2.5 * np.pi / 2),
-            CZ(2, 3),
-            PauliRotation((0, 1, 2), (3, 2, 1), 0.51 * np.pi / 2),
-        ]
-        circuit = QuantumCircuit(4, gate_list)
-        transpiled = CliffordApproximationTranspiler()(circuit)
-
-        expect = QuantumCircuit(4)
-        expect.extend(
-            [
-                H(0),  # H(0)
-                S(1),  # T(1)
-                SqrtY(0),  # RY
-                SqrtX(1),  # RX
-                S(1),  # U2
-                SqrtX(1),
-                Z(1),
-                Z(3),  # U3
-                SqrtX(3),
-                Z(3),
-                SqrtX(3),
-                Sdag(3),
-                CZ(2, 3),  # CZ
-                SqrtX(1),  # PauliRotation
-                H(2),
-                CNOT(1, 0),
-                CNOT(2, 0),
-                S(0),
-                CNOT(1, 0),
-                CNOT(2, 0),
-                SqrtXdag(1),
-                H(2),
             ]
         )
