+++ packages/cirq/tests/cirq/circuit/test_cirq_transpiler.py
@@ -8,7 +8,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from quri_parts.circuit import (
+from quri_parts.circuit.transpile import (
     PauliDecomposeTranspiler,
     PauliRotationDecomposeTranspiler,
     SequentialTranspiler,
